update components files structure
