+++ renderer/src/components/ActivitiesTable/index.ts
@@ -1,0 +1,1 @@
+export { default as ActivitiesTable } from './ActivitiesTable';
