+++ renderer/src/components/Bookings/index.ts
@@ -1,0 +1,1 @@
+export { default as Bookings } from './Bookings';
