+++ renderer/src/components/ConnectionsSection/ConnectionsSection.tsx
@@ -1,8 +1,8 @@
-import TrelloConnection from "../TrelloConnection";
-import GoogleConnection from "../GoogleConnection";
-import Office365Connection from "../Office365Connection";
-import JiraConnection from "../JiraConnection";
-import TimetrackerWebsiteConnection from "../TimetrackerWebsiteConncetion";
+import { TrelloConnection } from "../TrelloConnection";
+import { GoogleConnection } from "../GoogleConnection";
+import { Office365Connection } from "../Office365Connection";
+import { JiraConnection } from "../JiraConnection";
+import { TimetrackerWebsiteConnection } from "../TimetrackerWebsiteConncetion";
 
 const ConnectionsSection = () => (
   <section className="h-full">
