+++ renderer/src/components/DateSelector/index.ts
@@ -1,0 +1,1 @@
+export { default as DateSelector } from './DateSelector';
