+++ renderer/src/components/DateSelector/types.ts
@@ -1,0 +1,8 @@
+import { Dispatch, SetStateAction } from "react";
+
+export type DateSelectorProps = {
+  isDropboxConnected: boolean;
+  selectedDate: Date;
+  setSelectedDate: Dispatch<SetStateAction<Date>>;
+  selectedDateReport: string;
+};
