+++ renderer/src/components/FolderSelector/index.ts
@@ -1,0 +1,1 @@
+export { default as FolderSelector } from './FolderSelector';
