+++ renderer/src/components/FolderSelector/types.ts
@@ -1,0 +1,4 @@
+export type FolderSelectorProps = {
+  folderLocation?: string;
+  setFolderLocation: (location: string) => void;
+};
