+++ renderer/src/components/GoogleConnection/GoogleConnection.tsx
@@ -1,25 +1,14 @@
 import React, { useEffect, useState } from "react";
-import Button from "../shared/Button";
+import Button from "../../shared/Button";
 import { useRouter } from "next/router";
 import {
   getGoogleAuthUrl,
   getGoogleCredentials,
   getGoogleUserInfo,
-} from "../API/googleCalendarAPI";
+} from "../../API/googleCalendarAPI";
 import { ArrowRightOnRectangleIcon } from "@heroicons/react/24/solid";
 import isOnline from "is-online";
-
-type GoogleCredentails = {
-  access_token: string;
-  refresh_token: string;
-};
-
-export type GoogleUser = {
-  googleAccessToken: string;
-  googleRefreshToken: string;
-  userName: string;
-  accountId: string;
-};
+import { GoogleCredentails, GoogleUser } from './types';
 
 const GoogleConnection = () => {
   const router = useRouter();
