+++ renderer/src/components/GoogleConnection/index.ts
@@ -1,0 +1,1 @@
+export { default as GoogleConnection } from './GoogleConnection';
