+++ renderer/src/components/JiraConnection/index.ts
@@ -1,0 +1,1 @@
+export { default as JiraConnection } from './JiraConnection';
