+++ renderer/src/components/ManualInputForm/index.ts
@@ -1,0 +1,1 @@
+export { default as ManualInputForm } from './ManualInputForm';
