+++ renderer/src/components/Office365Connection/Office365Connection.tsx
@@ -1,32 +1,35 @@
 import React, { useEffect, useState } from "react";
 import { ArrowRightOnRectangleIcon } from "@heroicons/react/24/solid";
-import Button from "../shared/Button";
+import Button from "../../shared/Button";
 import isOnline from "is-online";
-import { JiraUser } from "../helpers/utils/jira";
-import { FlagIcon } from "@heroicons/react/24/outline";
+import { Office365User } from "../../helpers/utils/office365";
 
-const JiraConnection = () => {
+const Office365Connection = () => {
   const [users, setUsers] = useState(
-    JSON.parse(localStorage.getItem("jira-users")) || []
+    JSON.parse(localStorage.getItem("office365-users")) || []
   );
+  const [showEventsInTable, setShowEventsInTable] = useState(false);
 
   const handleSignInButton = async () => {
     const online = await isOnline();
 
     if (online) {
-      global.ipcRenderer.send("jira:login");
+      global.ipcRenderer.send("office365:login");
     } else {
       global.ipcRenderer.send("app:load-offline-page");
     }
   };
 
   const handleSignOutButton = (id: string) => {
-    const filteredUsers = users.filter((user: JiraUser) => user.userId !== id);
+    const filteredUsers = users.filter(
+      (user: Office365User) => user.userId !== id
+    );
 
     if (filteredUsers.length > 0) {
-      localStorage.setItem("jira-users", JSON.stringify(filteredUsers));
+      localStorage.setItem("office365-users", JSON.stringify(filteredUsers));
     } else {
-      localStorage.removeItem("jira-users");
+      localStorage.removeItem("office365-users");
+      localStorage.removeItem("showOffice365Events");
     }
 
     setUsers(filteredUsers);
@@ -37,20 +40,21 @@ const JiraConnection = () => {
     const authorizationCode = params.get("code");
 
     const { access_token, refresh_token } = await global.ipcRenderer.invoke(
-      "jira:get-tokens",
+      "office365:get-tokens",
       authorizationCode
     );
 
     if (!access_token) return;
 
-    const { account_id, email } = await global.ipcRenderer.invoke(
-      "jira:get-profile",
-      access_token
-    );
+    const { userPrincipalName, mail, displayName, id } =
+      await global.ipcRenderer.invoke(
+        "office365:get-profile-info",
+        access_token
+      );
 
-    const username = email || "";
+    const username = userPrincipalName || mail || displayName || "";
     const isUserExists = users.some(
-      (user: JiraUser) => account_id === user.userId
+      (user: Office365User) => id === user.userId
     );
 
     if (isUserExists) {
@@ -59,32 +63,48 @@ const JiraConnection = () => {
     }
 
     const user = {
-      userId: account_id,
+      userId: id,
       accessToken: access_token,
       refreshToken: refresh_token,
       username: username,
     };
-
+    global.ipcRenderer.send("send-analytics-data", "calendars_connections", {
+      calendar: "office365",
+    });
     const newUsers = [...users, user];
 
-    localStorage.setItem("jira-users", JSON.stringify(newUsers));
+    localStorage.setItem("office365-users", JSON.stringify(newUsers));
     setUsers(newUsers);
+  };
+
+  const handleCheckboxChange = () => {
+    localStorage.setItem(
+      "showOffice365Events",
+      (!showEventsInTable).toString()
+    );
+    setShowEventsInTable(!showEventsInTable);
   };
 
   useEffect(() => {
     if (
       window.location.search.includes("code") &&
-      window.location.search.includes("state=jiracode") &&
+      window.location.search.includes("state=office365code") &&
       !window.location.search.includes("error")
     ) {
       (async () => addUser())();
+    }
+
+    if (localStorage.getItem("showOffice365Events") === "true") {
+      setShowEventsInTable(true);
     }
   }, []);
 
   return (
     <div className="p-4 flex flex-col items-start justify-between gap-2 border rounded-lg shadow dark:border-dark-form-border">
       <div className="flex justify-between items-center w-full ">
-        <span className="font-medium dark:text-dark-heading">Jira</span>
+        <span className="font-medium dark:text-dark-heading">
+          Microsoft Office 365
+        </span>
         {!users.length && (
           <Button
             text="Add account"
@@ -92,7 +112,7 @@ const JiraConnection = () => {
             type="button"
           />
         )}
-        {/* {users.length > 0 && (
+        {users.length > 0 && (
           <button
             onClick={handleSignInButton}
             type="button"
@@ -102,7 +122,7 @@ const JiraConnection = () => {
               Add another account
             </span>
           </button>
-        )} */}
+        )}
       </div>
       <div className="flex items-center justify-between gap-4 w-full">
         {!users.length && (
@@ -131,19 +151,29 @@ const JiraConnection = () => {
         )}
       </div>
       <p className="text-sm text-gray-500  dark:text-dark-main">
-        After connection, you will be able to fill in the Description field with
-        tasks from the drop-down list. Just begin typing and in the drop-down
-        list you will see Jira tasks that starting with "JI::".
+        After connection, you will be able to fill in the Report with the
+        information from events of your Microsoft Outlook Calendar
         <br />
-        You can see only those tasks, where you are an assignee
-        <br />
-        <span className="flex gap-2 items-center">
-          <FlagIcon className="w-4 h-4" />
-          Ensure you have a Jira account connected to your Atlassian account.
-        </span>
+        You can authorize with a work, or personal Microsoft account (e.g.
+        Skype, Xbox)
       </p>
+      {users.length > 0 && (
+        <div className="flex items-start justify-between gap-6 w-full">
+          <p className=" max-w-sm text-sm text-gray-500 dark:text-dark-main">
+            Show Office365 events in activity table
+          </p>
+          <div>
+            <input
+              onChange={handleCheckboxChange}
+              className="w-4 h-4"
+              type="checkbox"
+              checked={showEventsInTable}
+            ></input>
+          </div>
+        </div>
+      )}
     </div>
   );
 };
 
-export default JiraConnection;
+export default Office365Connection;
