+++ renderer/src/components/Office365Connection/index.ts
@@ -1,0 +1,1 @@
+export { default as Office365Connection } from './Office365Connection';
