+++ renderer/src/components/SelectFolderPlaceholder/index.ts
@@ -1,0 +1,1 @@
+export { default as SelectFolderPlaceholder } from './SelectFolderPlaceholder';
