+++ renderer/src/components/SelectFolderPlaceholder/types.ts
@@ -1,0 +1,3 @@
+export type SelectFolderPlaceholderProps = {
+  setFolder: (path: string) => void;
+};
