+++ renderer/src/components/TimetrackerWebsiteConncetion/index.ts
@@ -1,0 +1,1 @@
+export { default as TimetrackerWebsiteConnection } from './TimetrackerWebsiteConncetion';
