+++ renderer/src/components/Totals/constants.ts
@@ -1,0 +1,5 @@
+export const TOTAL_PERIODS = [
+  { id: 0, name: "day" },
+  { id: 1, name: "week" },
+  { id: 2, name: "month" },
+];
