+++ renderer/src/components/Totals/index.ts
@@ -1,0 +1,1 @@
+export { default as Totals } from './Totals';
