+++ renderer/src/components/TrelloConnection/index.ts
@@ -1,0 +1,1 @@
+export { default as TrelloConnection } from './TrelloConnection';
