+++ renderer/src/components/UpdateDescription/index.ts
@@ -1,0 +1,1 @@
+export { default as UpdateDescription } from './UpdateDescription';
