+++ renderer/src/helpers/dates.ts
@@ -1,0 +1,1 @@
+export const DAY = 60 * 60 * 24 * 1000;
